Add in-memory store
